src/requirements.txt
# Add your requirements here

discord.py>=2.5.2
python-dotenv

